+++ app/widgets/Notifications/Notifications.php
@@ -52,7 +52,7 @@ class Notifications extends Base
                 'invite|' . $from,
                 $contact->truename,
                 $this->__('invitations.wants_to_talk', $contact->truename),
-                $contact->getPhoto(),
+                $contact->getPicture(),
                 4
             );
         }