+++ src/Movim/Console/CompileStickers.php
@@ -6,6 +6,7 @@
 
 namespace Movim\Console;
 
+use Movim\Image;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -21,7 +22,17 @@ class CompileStickers extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $count = compileStickers();
+        $count = 0;
+
+        foreach (glob(PUBLIC_PATH . '/stickers/*/*.png', GLOB_NOSORT) as $path) {
+            $key = basename($path, '.png');
+
+            if ($key != 'icon') {
+                $count++;
+                copy($path, PUBLIC_CACHE_PATH . hash(Image::$hash, $key) . '_o.png');
+            }
+        }
+
         $output->writeln('<info>'.$count.' stickers compiled</info>');
         return 0;
     }